fenrir: Update Pacman support to `BP2A.250605.031.A3`
Update patterns to add support for the latest NothingOS 4 stable
release on the Nothing Phone 2A.

Nothing patched the "vulnerability" in NothingOS 4 stable, but the
fix never reached the beta builds. Because of this it remains
possible to use fenrir on NothingOS 4 stable by downgrading the
Preloader to the NothingOS 4 beta version.

Remove payload injection for now, as it provides little value and is
time-consuming to keep porting across releases.